recountCategories.php: Automatically run cleanupEmptyCategories.php too
Instead of telling the operator to run another script, let's do it for
them. A --skip-cleanup option is provided to keep the current behavior.

We can use Maintenance::runChild() to set up the script, but because it
automatically passes options through to the child, we run into a
conflict that both scripts have a "mode" parameter with different
expected values. Calling loadParamsAndArgs() directly with empty options
and arguments works, except we need to modify the function to support
empty arrays. I checked all callers in Gerrit via Codesearch and none
should be affected.

Inspired by T299823#7643093.

Change-Id: I96303ce9b15b62b56494592e24d49154583dd269