Support audio on filepage in InstantCommons
The transform for ForeignAPIFile generally delegates transforms to the
thumburl of the remote. This works well, unless the file in question
has no inherent visual area, like audio files. Audio files on the file
description page when using InstantCommons have a 0x0 size, and the
Commons API returns errors for thumbnails of 0 size.

Instead, since we already know we don't need a thumbnail url for audio
fragments, just skip that step directly and simply pass an empty
string thumburl to the HTML generator of TMH.

Bug: T300085
Change-Id: Iea0e16de97e5e5da8c84bdbed8a2961ad194c040