Title: Set context page for `{{PAGENAME}}` etc. in getEditNotices()
The main call to Title::getEditNotices is in EditPage, where the global
title used by the message parser is correct. In other Action like
DeleteAction this would also be correct.

However, when called in API context (eg. ApiVisualEditor), or from a
special page that acts as a page action (such as
Special:EditMassMessageList or Special:Undelete) then the global title
is wrong (somewhat akin to wgPageName vs wgRelevantPageName).

Bug: T300184
Change-Id: I242e042317a1e16c8d51edbf7800c8b7d70d468e