resourceloader: Speed up mw.loader 404 test
The previous one required a live DNS lookup which was going over the
network through various layers until it timed out or got nothing.
Use a local 404 instead, which achieves the same result but returns
notably faster (3ms, instead of 50-150ms). Not exactly sure why
DNS is that slow, might have to do with it being in VM+container.

Change-Id: If3ef00b6ec744d4626d7cea2a481640b68af6f8e