HTMLForm: Hard-deprecate constructor without $context parameter
It was supposed to become required in 1.19, according to a code
comment (typo'ed in ead9055a).

There seem to be very few uses without the parameter out there (based
on a brief look at https://codesearch.wmcloud.org/), and most of them
are in tests, so they should be easy to find and correct.

Change-Id: I161cc342d1af813c281a6d9e30fdd85bc7b07578