resourceloader: Optimize hot code in makeQueryString()
In a devtools flamechart over a page view, I found that for the batch
request of page modules spends about 8ms (MacBook Pro, CPU 4x) just
figuring out the request URL. This is attributed to the line that
calls makeQueryString() in a loop to keep predicting the URL length.

I might change that more broadly, but for now we can cut it in half
by writing it more simply with a for-loop. This matches how similar
code in jQuery.param() and other libraries is written.

Micro bench at https://phabricator.wikimedia.org/P19931

```
makeQueryString({ lang, modules, skin, version})
i=10_000

// Firefox 96
// { before: 48ms, after: 24ms }

// Safari 15.2
// { before: 37ms, after: 13ms }

// Node.js v17.4.0
// { before: 37.837ms, after: 23.877ms }
```

Change-Id: Idf40fabf809f84bcbacf1a8e15adf77ab52397fa