Enable "ES6" modules in Safari 11+
Our current ES6 feature-detection code is perhaps overly-strict,
and excludes all versions of desktop and mobile Safari prior to 14.
This is causing features like MediaSearch to fail to run in places
where they ought to be supported
(see https://phabricator.wikimedia.org/T300094).

This change replaces the the check for non-BMP variable name support
with a check for arrow functions with default arguments. This is taken
from Benjamin De Cock's snippet on Github:
https://gist.github.com/bendc/d7f3dbc83d0f65ca0433caf90378cd95

Bug: T300666
Change-Id: I1cb370603b2ac13bcaf74890841ec941961810ef