Throw DBError if wrong database type has been chosen.
Currently it throws an exception when the database type has been changed
to unsupported one in localsettings.php.
The code displays a very confusing error in the browser when triggered,
to solve the error I used database exception and conditional statement.
If db type is not one of built-in supported or don't have custom
implementation DBerror will be thrown.

Bug: T212236
Change-Id: I5fce8b42a5e42e03a34aff5820f7b842393912f1