apisandbox: Mark internal and deprecated modules more prominently
Previously they were only marked as such in the detailed descriptions
hidden behind the [Expand] buttons, which no one clicks. (Deprecated
modules were also marked with a line-through in the dropdown list.)

Now they're also marked in the dropdown lists (which you have to
click) and in the descriptions (which you have to see).

Change-Id: Id19f0804c6b2ddf8d507561b5dfa69b15f7fd90f