skins: Memoize buildSidebar()
Skin::buildSidebar() accounts for 3.6% of index.php CPU time in
production, and local testing shows it being called 3 times for a
useskin=minerva page view. But it is stable for override, so we can't
implement caching transparently to the caller since subclasses would not
do caching.

There is SkinMinerva::buildSidebarCached() -- without that,
buildSidebar() would have been called 4 times instead of 3.

So, add a process cache for it.

Change-Id: Iffe63c465f6039336ebbdccd88f2a55fba8883d0