Make Database::ping() work if the connection was already lost
The pingAll() loop in updateSpecialPages.php tries salvage a
long-running process by waiting for a restarting DB server to come back
up again. The feature was broken because assertHasConnectionHandle()
would throw an exception if a previous reconnection attempt failed.

So, have ping() call replaceLostConnection() if the connection is
already missing.

Change-Id: I24b4cc18a7b586e014dcfd25b50653487fb18893