Generate abstract schemas with one script call
This rewrites both generateSchemaChangeSql.php and generateSchemaSql.php
to deduplicate logic and adds the 'all' option to --type to generate
schemas for each supported DBMS platform.

Specifying a path with --sql will have the script insert a directory
named after the platform, but only when --type=all is provided.
Only if a directory named mysql/ exists will sql files for mysql be
placed there, to allow using a setup similar to MediaWiki, where only
non-mysql types have their sql files placed in dedicated directories.

This also adds preparations for T298320, so that the abstract schema
can be validated and any errors can be handled before generating a
schema.

Bug: T268587
Change-Id: Ief8282017c8d38659b79262afb8fc691b5bda256