EditPage: make most methods private
One of the things that make refactoring EditPage even harder is the
large number of public/protected methods and the existence of
subclasses. This class changes the visibility to private for all methods
that are never used outside of EditPage itself.

I've manually checked codesearch (in "Everything") for usages of these
methods, and only changed the ones unused outside. A single usage, even
if in an unmaintained 3rd party extension, means that the method stays
protected/public for now. The only exception to this is ::tokenOk, which
is marked as @internal since I4c7c6e7b5a80e86456965521f88d1dfa7d698f84
(in 2020) and before that, as @private since [1] (in 2004); as such, any
external use of that method was ill-advised in the first place.

Although it happens quite commonly, this class is not stable to extend,
hence these changes do not need a deprecation phase.

Also replace a couple more deprecated methods and adjust some typehints.

[1] - https://gerrit.wikimedia.org/r/plugins/gitiles/mediawiki/core/+/7030f69a1ad8295facca0f4c9c5a3833cb703ecf

Bug: T252907
Change-Id: Ic5bd295cd896b05f0a88cb2085e11ad5b640e240
(cherry picked from commit a071d2c062f2cbff3ba7e0094f085de88805d6da)