ParserOutputAccess: Add process cache within the service class
This guards against duplicate parses.

These happen when a page is parsed but an extension needs the
ParserOutput again in the same request when it hasn't made into the
ParserCache yet, or if it is considered uncachable. In that case we
still want to allow re-use within the same process.

Bug: T301310
Change-Id: I1ddd967a40b760b1e53f1fd227cb0d0732f78ec1