Add script to benchmark TRUNCATE versus DELETE
I noticed that the CentralAuth integration tests were slow due to their
use of permanent rather than temporary tables, which cause MariaDB to do
unlink+fsync on truncate. The result was an incredible ~100ms of latency
per truncate query, reproducible with this script, although it was
reduced to ~10ms by mounting with nobarrier. Since this is likely
system-dependent, I am curious to know if others have the same issue.

Change-Id: Ifb25ff7675fff179e0b6e465aa3043b928048fc0