resourceloader: Pass exports parameter to package modules
This adds support for the `exports` alias for `module.exports`, but only
to package modules, per the discussion on T284511. This makes it easier
to use external libraries that use UMD, since UMD tests whether
typeof exports === "object", and if not, it adds the library to the
global scope instead of exporting it.

This is a breaking change for package modules that were expecting
exports to be undefined; in practice, this means modules that were
expecting UMD wrappers to add to the global scope have to be updated to
expect it to export the library instead. However, most UMD wrappers in
MW core are in non-package files, which are not affected by this change.

The affected libraries are updated accordingly:
- Codex used the global scope. This library is very new and nothing uses
  the global window.codex variable yet, so we don't need to provide
  backwards compatibility for it.
- Pako was patched to have its UMD wrapper look at module instead of
  exports; remove this patch since it's now unnecessary

Bug: T284511
Change-Id: Ie2e965cee001fb7f10b2b3e3f796d9511f6b63d2