mediawiki.action.styles: Bundle in categoryPage.less
This tiny stylesheet isn't worth keeping a separate bundle entrypoint
for. Document a basic budget for mediawiki.action.styles and move it
in there.

Given that category pages are cached in the CDN the same way as
article pages are, keep the old module around for a week as its cached
pages will still be loading that module, and those page did not load
the "mediawiki.action.styles" module yet.

Another way could have been to first add "mediawiki.action.styles"
to CategoryViewer.php, and have that load alongside the old module
for a week (without moving any styles). Then the week after, move
the styles over and delete the old module. The cached pages will
be referencing both and finding the new styles under the new module,
and the old module would simply be ignored as no longer existing.

Change-Id: If406ce2ae38eb165758add6099b2106794bf225c