Respect the wiki when performing autoblocks
Currently an autoblock is always inserted into the database of
the local wiki, even if the block is a cross-wiki one. This is
also the one of the issues that caused T281972.

In combination with Icdf9e4f37c3bfa568c712bebceccf9cc46db2ec8
this is going to fix the wrongly performed autoblocks by the
CheckUser extension.

I1051beedfd67c4e2c546b9eec04b6d592d969af3 is still a more clean
solution that this, but since that patch is blocked on T292375,
which is probably not going to be resolved in the near future,
this patch is chosen as a temporary solution.

Bug: T281972
Bug: T284873
Change-Id: I0d42c743ec188b20b636c8ec577d6cf02a7f1059