interwiki: Adjust documentation for InterwikiLookup::getAllPrefixes
Use bool instead of string, most caller provided int instead.

The InterwikiLookupAdapter class is strict comparing Interwiki::isLocal
(a bool) with the given $local value, which evaluates always to false
and the expected filter option is not working.
The unit tests is using bool, not the documented string.

The ClassicInterwikiLookup is not affected.

Found by phan strict checks

Change-Id: I63ed9772429c6f5bfe38a9fad190491812866f3f