ApiDelete: add option to delete associated talk page
This works thanks to the new "deletetalk" parameter (the name is
consistent with "movetalk" to ApiMove). The API will return with a
warning if the page is already a talk page, or if the talk page
doesn't exist.

Note that the API response doesn't expose information about the talk
page deletion. This would likely be a breaking change to the module, and
is also not too easy to implement until FileDeleteForm is refactored as
well.

Bug: T263209
Change-Id: I30f2da190433a3ef9c0ac375f23b68acd3d6f190