Push down date header code from IndexPager to ReverseChronologicalPager
Follow-up to ed63ede6575056e1 (Iaa55eab57f76), I didn't think
of this in code review.

In IndexPager, we're not necessarily paging by timestamp, and the
things we're paging might not even *have* timestamps (e.g. consider
CategoryPager used on Special:Categories), so it seems wrong to have
this code there.

(Maybe some kind of headers could be useful, but there's currently no
such use case, and the could would need to be made a bit more abstract
anyway.)

Change-Id: I5263375f0d9dd98fd37c40a1333f4154c2b41e98