Allow LocalSettings.php to be loaded in function scope.
This adds an experimental mode that loads LocalSettings.php in a
function, rather than in top level global scope. This is controlled
using the MW_USE_LOCAL_SETTINGS_LOADER environment variable.

This is not intended for use in production as this time.
The experimental mode should be available in the 1.38 release,
so people can try it out with different setups.

Bug: T304183
Depends-On: Ie6245ff8cd2bc7bab2af3b1450070185dbc2d824
Change-Id: I0f20024803336064ecb07b51fa56581af7b67a85
(cherry picked from commit 33a3c3c68791929b4cf367a212db3bb9a78ac64f)