resourceloader: Fix jsminplus regex recursion
The outer repeated subexpression in the regexes for string literals and
regex literals was consuming stack space, leading to
PREG_JIT_STACKLIMIT_ERROR for strings with many escape sequences, or
long regex literals of any kind (>8190 bytes in my setup).

So, use a possessive quantifier for the three affected loops. This
should be safe as long as the next part of the sequence (the end
delimiter) cannot match the loop body. So for example /".*"/ would be
unsafe for this optimisation, whereas /"[^"]*"/ would be safe. My
analysis is that these regexes are like the latter case.

Bug: T299537
Change-Id: Ic00d21fe96d4a6adec33dc3ca0d64c032af1ffe3