Skin: Streamline confusing "checkTitle" related code
This is a direct follow up for what was done in I87c76fd.

Notable:

* I removed multiple "this is an invalid title" titles. These are
effectively debug messages, but why should the user see them? Either
throw an exception, or render a link that is clearly broken – as I
suggest here.

* "checkTitle" is a helper method with a strong side effect that is
not obvious from the name, in my opinion. I could not find relevant
usages and removed it.
https://codesearch.wmcloud.org/search/?q=:checkTitle%5C(&files=%5C.php%24

I propose to remove the method without deprecation because it is very
obviously meant to be a private helper method. The way it's designed
(especially in this rather specific class) it doesn't make sense to
call it from anywhere else.

Depends-On: Ic7f168fa89f0788ca679eac067b40e735f3dcc46
Change-Id: Ic776a6cdc6be30feb92b7224ef13994c14db726c