Trivial replacements of "-" comparisons with isDisabled()
This is split from I1d267ae, as well as a follow-up to Ie70d849.

Custom code like this exists because some messages do something
like {{#if: … | … | - }} to get dynamically disabled. This is not
something isDisabled() supports. That acts on the unparsed message
before any {{…}} was expaned.

The few cases in this patch are not affected because they use plain(),
which does not do any {{…}} expansion anyway.

Change-Id: I70541cf28c6fff14422e202f774b86086c2a850c