HistoryPager: Refactor to avoid special handling for the last row
Displaying each row requires the revision ID from the next row, to
create the 'undo' and 'prev' links. (These links specifically refer to
the next displayed row, according to the current filters, and not just
the previous revision, like the byte diff numbers.)

Previously this class would override normal row formatting to display
the previous row instead when the normal pager would display the
current row, and then handled the last row specially. This code dates
back to 2006: 20949b0961. It was complicated, and got more complicated
when grouping by dates was introduced (ed63ede657), when it caused the
last row to be displayed in a separate list.

Thanks to the addition of batch processing in f7f84dddb3, we don't
need to do that any more: the necessary information about revisions is
pre-computed and we can just look up the ID for next row while
displaying the previous one.

Bug: T303210
Change-Id: I838b0bd4c04faf4f83b21282deb757207ab1732a