api-testing: Remove 'matched_title' check
Sqlite and MySQL are returning results in different
orders for the REST Search handler, which affects how it handles
de-duplicating results with redirects.

If the Redirect Target is processed first, the matched_title
field is not populated because that page is not a redirect.
If the Redirect Source is processed first, the matched_title
field IS populated because that page is a redirect.

Either way, we don't have duplicate results (which is the most
important part). Until the logic is consistent, remove the
matched_title check.

Bug: T302706
Change-Id: Ic3977655565aa9f6d6c184749706273b0315b7be