Try to ensure tags are arrays
In a similar vein to I46033d5ddda055c275dbda71d5c38c4a7a858c9d

The documentation to WikiPage::doUpdateRestrictions is changed to remove
`|string|null`.  ManualLogEntry::addTags can handle a string but it's
unexpected in the RevisionFromEditComplete hook.  From codesearch,
ApiProtect is the only caller passing in $tags, and that's coming from
a PARAM_ISMULTI, which presumably can only be an array or null.

Bug: T302918
Change-Id: Icd8a7391f7a8c21b0cdb1f734e65ee5661c1075f