In wfArrayInsertAfter(), skip insert on missing key
Instead of raising a notice due to $offsetByKey[$after] referring to a
nonexistent array element, if the item referred to by $after is not
present, don't insert the array.

This is DWIM based on code review of the callers. wfArrayInsertAfter()
is mostly used to extend skin data. BetaFeatures, Favorites,
LiquidThreads and WebToolsManager are inserting an element after
preferences/watchlist only if the user is logged in, assuming that
logged in users have preferences. Temporary users are logged in
without preferences, breaking this assumption, but in all cases it is
acceptable to omit the link from the skin for such users.

Bug: T300263
Change-Id: I76473ef668ca4d9200ca42b57d793e1e03799ba6