EditPage: Avoid parsing wikitext in comment
Wrapping the comment in <nowiki> can be bypassed if the comment itself
contains </nowiki>, which though usually pointless is allowed. (But this
bypass only enables the comment to be parsed as wikitext, not directly
used as HTML, so it’s not a security issue as far as I can tell.)
Properly inform Message that it’s supposed to be a plain-text parameter
instead. (Using wfEscapeWikiText( $comment ) would also work, but
plaintextParams() seems cleaner to me. Note that the user name must be a
regular param, though.)

Bug: T303116
Change-Id: Id89075c91eda41a3ebc8b3baf733c0cbefb6275c