Use Message::newFromSpecifier() in wfMessage()
As a result, when an existing Message object is passed in the $key
parameter, the function will return a clone of that object, rather
than a new Message constructed based on the key extracted from the
old message.

This is mostly relevant when the passed object is an instance of
a subclass like RawMessage or ApiMessage, but it could have other
unpredictable consequences. No test failures, though!

Change-Id: I4815e92dfaf5fbcc8c9356106738716c6b9ae0da